feat: finalização do módulo do professor (Telas 3, 4, 5 e 6)

+++ lib/data/modelos/modelo_registro_aluno.dart
@@ -2,7 +2,7 @@ class ModeloRegistroAluno {
   final int? id;
   final int alunoId;
   final String disciplina;
-  final double? nota;
+  final int? nota;
   final int? faltas;
 
   ModeloRegistroAluno({
